fix: remove dead code
